dd: add owner to meta-data
Use 'nobody' user to mark dump directories which have no owner and are
publicly readable and chownable. The 'nobody' should not be used
directly but through the function set_no_owner() which is a substitution
for adding S_IROTH (man 2 open) to dump dir mode.

Use meta-data to get the owner only if the dump directory is owned by
super-user, because it is the only state when we can trust that value.

Related: #354

Signed-off-by: Jakub Filak <jfilak@redhat.com>