problem_data: introduce file descriptor items
Add file_descriptor member to struct problem_item. This member can be
used by all problem_item types. If its value is not
PROBLEM_ITEM_UNINITIALIZED_FD, the value is a valid fd of the element file.

Introduce CD_FLAG_FD. If an item has this flag, the item's
file_descriptor must be a valid file descriptor and the item's content
is ('/proc/self/fd/%d'), which is valid file, so the backward
compatibility with clients supporting only CD_FLAG_BIN is not broken.
Items having this flag up must have CD_FLAG_BIN up too.

Obtaining size of these items is not supported because, it would require
reading of the file descriptor to EOF and we know nothing about the
descriptor (it could be a stream).

Introduce problem_data_add_file_descriptor() which takes care of adding
file descriptors to problem_data.

Required for Problems2 D-Bus API.

Signed-off-by: Jakub Filak <jfilak@redhat.com>