wizard: correct comments in save_text_if_changed()
The FIXME has become irrelevant in commit
257578a23d1537a2d235aaa2b1488ee4f818e360 because there are no more
problems with other widgets as the function no longer reloads data from
the dump directory.

I added a comment explaining how the wizard deals with errors while
saving user data, because it is not clear and one might think that the
errors are silently ignored.

Signed-off-by: Jakub Filak <jfilak@redhat.com>