mailx: stop creating dead.letter on mailx failures
SELinux does not like mailx creating the file and I do not see any
reason to create the file in a problem directory because the file
contains a copy of the email that could not be send.

Failures of EVENT=notify are discoverable in system logs and if you run
the reporter manually, you will see that mailx failed.

Signed-off-by: Jakub Filak <jfilak@redhat.com>